fix: improved folder structure
