+++ packages/core/src/app.module.ts
@@ -61,8 +61,8 @@ import { TenantModule } from './tenant/tenant.module';
 import { EmailTemplateModule } from './email-template/email-template.module';
 import { EquipmentModule } from './equipment/equipment.module';
 import { EmployeeLevelModule } from './organization_employee-level/organization-employee-level.module';
-import { ExportAllModule } from './export_import/export-all.module';
-import { ImportAllModule } from './export_import/import/import-all.module';
+import { ExportAllModule } from './export-import/export/export-all.module';
+import { ImportAllModule } from './export-import/import/import-all.module';
 import { TaskModule } from './tasks/task.module';
 import { EquipmentSharingModule } from './equipment-sharing/equipment-sharing.module';
 import { OrganizationEmploymentTypeModule } from './organization-employment-type/organization-employment-type.module';
