+++ packages/core/src/export-import/export/export-all.controller.ts
@@ -7,9 +7,9 @@ import {
 	Query
 } from '@nestjs/common';
 import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
-import { ExportAllService } from './export-all.service';
 import { AuthGuard } from '@nestjs/passport';
-import { ParseJsonPipe } from './../shared/pipes/parse-json.pipe';
+import { ExportAllService } from './export-all.service';
+import { ParseJsonPipe } from './../../shared/pipes/parse-json.pipe';
 
 @ApiTags('Download')
 @UseGuards(AuthGuard('jwt'))
