+++ packages/core/src/export-import/export/export-all.service.ts
@@ -112,7 +112,7 @@ import {
 	TimeSlot,
 	User,
 	UserOrganization
-} from './../core/entities/internal';
+} from './../../core/entities/internal';
 
 export interface IRepositoryModel {
 	repository: any;
